add: dataset id for backup

+++ client/models.go
@@ -42,7 +42,7 @@ type StoreSourceFileResponse struct {
 	Result  struct {
 		Code    string `json:"code"`
 		Message string `json:"message,omitempty"`
-		Data    int    `json:"data,omitempty"`
+		Data    int64  `json:"data,omitempty"`
 	} `json:"result"`
 	Id int `json:"id"`
 }
